Move interaction handling code to each command's folder
