+++ src/Commands/Expansions/DisableExpansionCommand.ts
@@ -1,7 +1,16 @@
-import {Expansion, displayName} from "../../Civs/Expansions";
+import {Expansion, displayName, stringToExpansion} from "../../Civs/Expansions";
 import { loadUserData, saveUserData } from "../../UserDataStore";
+import {ChatInputCommandInteraction} from "discord.js";
 
-export async function disableExpansionCommand(tenantId: string, expansion: Expansion): Promise<string> {
+export async function handleDisableExpansion(interaction: ChatInputCommandInteraction) {
+    const serverId = interaction.guildId!;
+    const expansion = stringToExpansion(interaction.options.getString("expansion")!)!;
+
+    const message = await disableExpansion(serverId, expansion);
+    await interaction.reply(message);
+}
+
+async function disableExpansion(tenantId: string, expansion: Expansion): Promise<string> {
     const userData = await loadUserData(tenantId);
 
     const defaultDraftSettings = userData.userSettings[userData.game].defaultDraftSettings;
