+++ src/Commands/Expansions/EnableExpansion.ts
@@ -1,8 +1,16 @@
-import {Expansion, displayName} from "../../Civs/Expansions";
+import {Expansion, displayName, stringToExpansion} from "../../Civs/Expansions";
 import { loadUserData, saveUserData } from "../../UserDataStore";
-import {User} from "discord.js";
+import {ChatInputCommandInteraction, User} from "discord.js";
 
-export async function enableExpansionCommand(tenantId: string, expansion: Expansion): Promise<string> {
+export async function handleEnableExpansion(interaction: ChatInputCommandInteraction) {
+    const serverId = interaction.guildId!;
+    const expansion = stringToExpansion(interaction.options.getString("expansion")!)!;
+
+    const message = await enableExpansion(serverId, expansion);
+    await interaction.reply(message);
+}
+
+async function enableExpansion(tenantId: string, expansion: Expansion): Promise<string> {
     const userData = await loadUserData(tenantId);
 
     const defaultDraftSettings = userData.userSettings[userData.game].defaultDraftSettings;
